refactor code into separate modules
